+++ src/main/java/com/estapar/garage/repository/ParkingSpotRepository.java
@@ -3,6 +3,7 @@ package com.estapar.garage.repository;
 import com.estapar.garage.entity.ParkingSpot;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -10,5 +11,5 @@ public interface ParkingSpotRepository extends JpaRepository<ParkingSpot, Long> 
     long countBySector(String sector);
     long countBySectorAndOccupiedTrue(String sector);
 
-    Optional<ParkingSpot> findByLatAndLng(Double lat, Double lng);
+    Optional<ParkingSpot> findByLatAndLng(BigDecimal lat, BigDecimal lng);
 }
